requirements.txt
SpeechRecognition
pyttsx3
google-generativeai
pyaudio
